Clear metric before setting new values (#23)
When a vulnerability is ignored, getting upgradeable etc. we get a new
label combination leading to duplicated metrics.

This change clears the gauge before setting newly scraped values. A 
mutex is added to protect the /metrics endpoint while setting new
values. This should avoid weird scrape values in case of prometheus 
hitting the exporter at the same time as a scraping is done.

Closes #21