Increase default interval to 10 minutes
Currently the scrape interval is 1 minute. As the metrics we are tracking are
rather slow moving we can accept a delay of up to 10 minutes for changes to be
reflected. This reduces the load in Snyk APIs significantly and also reduces
resource usage of the exporter (as it does the scraping 10 times less).