Migrate to aggregated issues
Snyk has deprecated the issues API endpoint in favour of aggregated-issues. The
new endpoint should be more precise and also does some of the logic currently
done in the exporter, ie. deduplication. It also reports vulnerabilities and
license issues as the same entity, ie. issue, which has the positive side effect
that the exporter now tracks license issues as well.

This change updated the endpoint and DTOs and removes the deduplication logic.

The aggregateIssues test is updated to use the new models and naming is aligned
to issues rather than vulnerabilities.